Harden relay init caching, gateway teardown order, and build guards

+++ gateway/include/roqr/gateway/gap.hpp
@@ -6,8 +6,8 @@
 
 namespace roqr::gateway {
 
-// Per-flow video continuity tracker (draft s8). Feed each video frame's
-// timestamp and its MediaClass; returns whether to deliver it. On a
+// Tracks continuity for a single video timeline (draft s8). Feed each video
+// frame's timestamp and its MediaClass; returns whether to deliver it. On a
 // suspected gap (timestamp regression or a jump past kJumpThreshold) it
 // drops non-recovery frames until the next keyframe or sequence header.
 class GapTracker {
@@ -19,7 +19,9 @@ public:
                              cls == roqr::rtmp::MediaClass::SequenceHeader;
         if (have_) {
             const bool regressed = timestamp < last_ts_;
-            const bool jumped = timestamp > last_ts_ + kJumpThreshold;
+            const bool jumped =
+                timestamp >
+                static_cast<uint64_t>(last_ts_) + kJumpThreshold;
             if (regressed || jumped) discontinuous_ = true;
         }
         have_ = true;
